requirements.txt
streamlit==1.39.0
tensorflow==2.15.0
opencv-python-headless==4.10.0.84
numpy
pillow
matplotlib
